refactor: Modal 작업에 대한 불필요 코드 제거 및 구조 수정 (#26)
* chore: 불필요 예시 페이지 제거

* refactor: `ModalContext` providers로 분리

* refactor: `@types`를 `src/types`로 분리

* chore: import 수정

* refactor: 전역 타입 폴더 `@types` 재차 분리

* chore: `Modal`에 대한 경로 수정

* `components/modal`에서 `components/common/Modal`로 수정
