--- src/components/modal/index.ts
@@ -1,11 +1,0 @@
-import ModalContent from '@/components/modal/ModalContent';
-import ModalRoot from '@/components/modal/ModalRoot';
-import ModalToggle from '@/components/modal/ModalToggle';
-import ModalPortal from '@/components/modal/Portal';
-
-export const Modal = {
-  Root: ModalRoot,
-  Content: ModalContent,
-  Toggle: ModalToggle,
-  Portal: ModalPortal,
-};
