+++ src/types/modal.d.ts
@@ -1,0 +1,9 @@
+import { Dispatch } from 'react';
+
+export type ModalActionType = 'OPEN' | 'CLOSE';
+
+export type ModalAction = { type: ModalActionType };
+
+export type ModalState = { isOpen: boolean };
+
+export type ModalDispatch = Dispatch<ModalAction>;
