--- src/types/modal.ts
@@ -1,3 +1,0 @@
-export type ModalActionType = 'OPEN' | 'CLOSE';
-
-export type ModalAction = { type: ModalActionType };
